test: migrate shell specs to main process
